Added the rest of the refactoring suggestion
