+++ backend/api/schemas.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+class StationSchema(BaseModel):
+    name: str
+    latitude: float
+    longitude: float
+    power: float
