+++ backend/db/mongo.py
@@ -1,0 +1,10 @@
+from motor.motor_asyncio import AsyncIOMotorClient
+from bson.objectid import ObjectId
+import os
+
+MONGO_URL = os.getenv("MONGO_URL", "mongodb://localhost:27017")
+
+client = AsyncIOMotorClient(MONGO_URL)
+db = client["berlin_bezirke_db"]
+
+station_collection = db.get_collection("charging_stations")
