--- berlin_postleitzahlen/berlin_postleitzahlen.cpg
@@ -1,1 +1,0 @@
-UTF-8
