+++ docs/installation_guide.txt
@@ -1,0 +1,29 @@
+# -------------------------------
+# Installation:
+# install Python 3.10+ on your system, restart system
+# Example here: Windows 11
+
+# -------------------------------
+# Installation:
+c:
+cd C:\(...)\src
+python -m venv .venv
+.venv\Scripts\activate.bat
+pip install -r requirements.txt
+pip install spyder
+
+cd C:\(...)\src
+
+# -------------------------------
+# Launch:
+c:
+cd C:\(...)\src
+.venv\Scripts\activate.bat
+spyder
+
+# jupyter-notebook
+
+
+# -------------------------------
+# Start Streamlit App:
+streamlit run main.py
