+++ requirements.txt
@@ -13,5 +13,11 @@ Folium
 openpyxl
 FastAPI
 pytest==8.3.4
+pymongo
+motor
+fastapi
+uvicorn
+requests
+dotenv
 #install local packages which are recognizable by the __init__.py file
 -e .
