--- src/tests/test_charging_station_rating/test_charging_station_rating_service.py
@@ -1,22 +1,0 @@
-"""
-test_charging_station_rating_service.py
-
-Description:
-    This script contains unit tests for the methods of the charging station rating service in 
-    src/charging_station_rating/charging_station_rating_service.py.
-
-Usage:
-    It is run with pytest
-
-Dependencies:
-    pytest
-
-Author:
-    Nina Immenroth
-
-Date:
-    2024-12-25
-
-Version:
-    Version 0.1
-"""
